+++ src/main/java/com/leoman/team/service/impl/TeamServiceImpl.java
@@ -121,7 +121,7 @@ public class TeamServiceImpl extends GenericManagerImpl<Team, TeamDao> implement
                     list.add(cb.like(root.get("name").as(String.class), "%" + t.getName() + "%"));
                 }
                 if(t.getCity().getCityId() != null){
-                    list.add(cb.equal(root.get("city").get("cityId").as(Integer.class), t.getCity().getCityId() ));
+                    list.add(cb.equal(root.get("city").get("cityId").as(Long.class), t.getCity().getCityId() ));
                 }
 
                 Predicate[] p = new Predicate[list.size()];
